server/requirements.txt
azure-identity
openai
httpx
os
PIL
json
python-dotenv



